Reorganize reviews module files
